⚡️ `lru_cache` table builds
I didn't benchmark it, but this might provide a small improvement if
there are concurrent users. Instead of performing an inner join on two
columns each time a page is built, this will compute the join on the
first request and cache it for subsequent use.

I *suspect* this is small for the current size of the user table.